requirements.txt
asn1crypto==0.24.0
boto==2.49.0
boto3==1.9.182
botocore==1.12.182
bz2file==0.98
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cycler==0.10.0
docutils==0.14
gensim==3.4.0
idna==2.8
jaconv==0.2.4
jmespath==0.9.4
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.0
mecab-python3==0.996.2
numpy==1.17.2
pandas==0.24.2
pycparser==2.19
pydub==0.23.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
smart-open==1.8.4
tornado==6.0.2
tqdm==4.32.1
urllib3==1.24.3
